fix: shader/CPU header compatibility and ResourceBindingMgr layout lifetime
- Guard MAX_LIGHT_NUM array in constants.h under __cplusplus to avoid GLSL redefinition
- Fix shader_structs.h include path (absolute -> relative)
- Use #ifdef __cplusplus to keep LightType as enum in C++ and #define in GLSL
- Change light_num field from ushort[] to uint[] for std140 alignment
- Persist glob_desc_set_layout_ as member to avoid dangling reference in requestDescriptorSet

+++ source/engine/functional/world/world.cpp
@@ -116,8 +116,8 @@ void World::loadedMesh2World() {
     for (auto light_type = 0; light_type < LightType::LIGHT_TYPE_NUM;
          ++light_type) {
       auto cp_num =
-          std::min(static_cast<ushort>(MAX_LIGHT_NUM[light_type] -
-                                       lighting_.light_num[light_type]),
+          std::min(static_cast<uint>(MAX_LIGHT_NUM[light_type]) -
+                       lighting_.light_num[light_type],
                    scene_data.lighting.light_num[light_type]);
       memcpy(ptr[light_type], scene_data.lighting.directional_lights,
              cp_size[light_type] * cp_num);
